requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
dj-rest-auth==7.0.1
Django==5.1.5
django-allauth==65.3.1
django-dotenv==1.4.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-http-client==3.3.7
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
